# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StreamVault API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /watch-history:
    get:
      tags: [watchHistory]
      operationId: getWatchHistory
      summary: Get the current user's watch history
      responses:
        "200":
          description: Watch history list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchHistoryList"
        "401":
          description: Unauthorized
    post:
      tags: [watchHistory]
      operationId: addWatchHistory
      summary: Record a watched item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddWatchHistoryInput"
      responses:
        "200":
          description: Recorded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
  /watch-history/{mediaId}:
    delete:
      tags: [watchHistory]
      operationId: deleteWatchHistoryItem
      summary: Remove an item from watch history
      parameters:
        - name: mediaId
          in: path
          required: true
          schema:
            type: integer
        - name: mediaType
          in: query
          required: true
          schema:
            type: string
            enum: [movie, tv]
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
components:
  schemas:
...
    WatchHistoryItem:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        mediaId:
          type: integer
        mediaType:
          type: string
        title:
          type: string
        posterPath:
          type: string
          nullable: true
        backdropPath:
          type: string
          nullable: true
        season:
          type: integer
          nullable: true
        episode:
          type: integer
          nullable: true
        episodeName:
          type: string
          nullable: true
        watchedAt:
          type: string
      required:
        - id
        - userId
        - mediaId
        - mediaType
        - title
        - watchedAt
    WatchHistoryList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WatchHistoryItem"
      required:
        - items
    AddWatchHistoryInput:
      type: object
      properties:
        mediaId:
          type: integer
        mediaType:
          type: string
          enum: [movie, tv]
        title:
          type: string
        posterPath:
          type: string
          nullable: true
        backdropPath:
          type: string
          nullable: true
        season:
          type: integer
          nullable: true
        episode:
          type: integer
          nullable: true
        episodeName:
          type: string
          nullable: true
      required:
        - mediaId
        - mediaType
        - title
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success